refactor: reorganize settings subfolders
