+++ src/renderer/src/pages/settings/appearance/AppearanceSettings.tsx
@@ -1,3 +1,3 @@
-import { ThemeSection } from '../ThemeSection';
+import { ThemeSection } from './ThemeSection';
 
 export const AppearanceSettings = () => <ThemeSection />;
