+++ src/renderer/src/pages/settings/appearance/ThemeSection.tsx
@@ -3,11 +3,15 @@ import { useContext } from 'react';
 import {
   ThemeContext,
   themes,
-} from '../../../../modules/personalization/browser';
-import { MoonIcon, SunIcon } from '../../components/icons';
-import { Label } from '../../components/inputs/Fieldset';
-import { Radio, RadioField, RadioGroup } from '../../components/inputs/Radio';
-import { SectionHeader } from '../shared/settings/SectionHeader';
+} from '../../../../../modules/personalization/browser';
+import { MoonIcon, SunIcon } from '../../../components/icons';
+import { Label } from '../../../components/inputs/Fieldset';
+import {
+  Radio,
+  RadioField,
+  RadioGroup,
+} from '../../../components/inputs/Radio';
+import { SectionHeader } from '../../shared/settings/SectionHeader';
 
 export const ThemeSection = () => {
   const { theme, resolvedTheme, setTheme } = useContext(ThemeContext);
