+++ src/renderer/src/pages/settings/appearance/index.ts
@@ -1,0 +1,1 @@
+export { AppearanceSettings } from './AppearanceSettings';
