+++ src/renderer/src/pages/settings/general/index.ts
@@ -1,0 +1,1 @@
+export { GeneralSettings } from './GeneralSettings';
