+++ src/renderer/src/pages/settings/index.ts
@@ -1,2 +1,4 @@
-export { AppearanceSettings, GeneralSettings, SyncSettings } from './tabs';
+export { AppearanceSettings } from './appearance';
+export { GeneralSettings } from './general';
 export { Settings } from './Settings';
+export { SyncSettings } from './sync';
