+++ src/renderer/src/pages/settings/sync/SyncSettings.tsx
@@ -1,0 +1,3 @@
+import { SyncProvidersSection } from './SyncProvidersSection';
+
+export const SyncSettings = () => <SyncProvidersSection />;
