+++ src/renderer/src/pages/settings/sync/index.ts
@@ -1,0 +1,1 @@
+export { SyncSettings } from './SyncSettings';
