--- src/renderer/src/pages/settings/tabs/SyncSettings.tsx
@@ -1,3 +1,0 @@
-import { SyncProvidersSection } from '../sync-providers/SyncProvidersSection';
-
-export const SyncSettings = () => <SyncProvidersSection />;
