--- src/renderer/src/pages/settings/tabs/index.ts
@@ -1,3 +1,0 @@
-export { AppearanceSettings } from './AppearanceSettings';
-export { GeneralSettings } from './GeneralSettings';
-export { SyncSettings } from './SyncSettings';
